requirements.txt
flask
requests
beautifulsoup4
google-generativeai
python-dotenv
gunicorn
flask-sqlalchemy
